fix/donation-request-form (#56)
* fix: able to proceedduring edit if all checkboxes are unselected

* fix: date time pre selected value

* fix: changed text for donation event page

* fix: omitPoints checkbox during creation

* fix: seed data regarding s3 images

* fix: data source merge conflicts

+++ client/src/pages/Donor/DonationRequestForm.tsx
@@ -49,11 +49,11 @@ export default function DonationRequestForm() {
     useState<dataToDonationRequestFormType | null>(null);
   const [donationRequest, setDonationRequest] = useState<DonationRequestType>({
     donationRequestId: 0,
-    donationEventId: 0, 
+    donationEventId: 0,
     dropOffDate: new Date(),
     dropOffTime: "",
-    omitPoints: true,
-    submittedBy: 0, 
+    omitPoints: false,
+    submittedBy: 0,
     oldDonationRequestItems: [],
     newDonationRequestItems: [],
   });
@@ -87,7 +87,7 @@ export default function DonationRequestForm() {
     if ("omitPoints" in updatedCheckedState) {
       setDonationRequest((prevData) => ({
         ...prevData,
-        omitPoints: !prevData.omitPoints,
+        omitPoints: !updatedCheckedState.omitPoints,
       }));
     } else {
       let updatedSelectedItems: Record<string, string | number>[] = [];
@@ -163,12 +163,11 @@ export default function DonationRequestForm() {
   };
 
   const handleDateTimeChange = (dateTime: Dayjs | null) => {
-    if (dateTime !== null) {
-      const formattedDate = dayjs(dateTime).format('DD/MM/YYYY');
+    if (dateTime !== null) { // TODO: don't need to store a time attribute in db
       const formattedTime = dayjs(dateTime).format('HH:mm');
       setDonationRequest((prevData) => ({
         ...prevData,
-        dropOffDate: dayjs(formattedDate, 'DD/MM/YYYY').toDate(),
+        dropOffDate: dateTime.toDate(),
         dropOffTime: formattedTime,
       }));
     }
@@ -213,32 +212,36 @@ export default function DonationRequestForm() {
             );
 
             if (!foundItem) {
-              console.log(donationRequestItem.id);
               deleteDonationRequestItemIds.push(donationRequestItem.id);
             }
           }
         );
-
-        Promise.all([
-          handleDeleteDonationRequestItem(deleteDonationRequestItemIds),
-          handleCreateDonationRequest(donationRequest),
-          handleUpdateDonationRequest(donationRequest),
-        ])
-          .then(([deleteStatus, createStatus, updateStatus]) => {
-            if (
-              deleteStatus.every((status) => status) &&
-              createStatus &&
-              updateStatus
-            ) {
-              console.log(status);
-              navigate("/donation-requests");
-            } else {
+        
+        if (
+          deleteDonationRequestItemIds.length !==
+          donationRequest.oldDonationRequestItems?.length ||
+          donationRequest.newDonationRequestItems.length >= 1
+        ) {
+          Promise.all([
+            handleDeleteDonationRequestItem(deleteDonationRequestItemIds),
+            handleCreateDonationRequest(donationRequest),
+            handleUpdateDonationRequest(donationRequest),
+          ])
+            .then(([deleteStatus, createStatus, updateStatus]) => {
+              if (
+                deleteStatus.every((status) => status) &&
+                createStatus &&
+                updateStatus
+              ) {
+                navigate("/donation-requests");
+              } else {
+                setError(true);
+              }
+            })
+            .catch(() => {
               setError(true);
-            }
-          })
-          .catch(() => {
-            setError(true);
-          });
+            });
+        }
       }
     }
   };
@@ -310,7 +313,7 @@ export default function DonationRequestForm() {
         donationRequestId: formData.donationRequestId || 0,
         dropOffDate: formData.dropOffDate || new Date(),
         dropOffTime: formData.dropOffTime || "",
-        omitPoints: formData.omitPoints || true,
+        omitPoints: formData.omitPoints || false,
         submittedBy: Number(localStorage.getItem("ecoyah-id")) || 0,
         oldDonationRequestItems: formData.donationRequestItems || [],
       }));
@@ -404,7 +407,7 @@ export default function DonationRequestForm() {
               name: "Receive Cashback Upon A Successful Donation",
               value:
                 action === "edit"
-                  ? !donationEventInfo.omitPoints ?? false
+                  ? !donationRequest.omitPoints
                   : false,
             },
           ]}
